Stats menu done
