+++ Assets/Scripts/Rover/RoverAIController.cs
