+++ Assets/Scripts/Rover/RoverAIController.cs.meta
