+++ Assets/Scripts/Stats.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 2520d277a636b4f4da778908a02fa62c
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
