+++ Assets/Scripts/Stats/StatsPanel.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 53a945c8d1ab9784889459b5fb8be876
